+++ src/spec-node/devContainers.ts
@@ -8,11 +8,10 @@ import * as crypto from 'crypto';
 import * as os from 'os';
 
 import { DockerResolverParameters, DevContainerAuthority, UpdateRemoteUserUIDDefault, BindMountConsistency, getCacheFolder } from './utils';
-import { createNullPostCreate, finishBackgroundTasks, ResolverParameters, UserEnvProbe } from '../spec-common/injectHeadless';
+import { createNullLifecycleHook, finishBackgroundTasks, ResolverParameters, UserEnvProbe } from '../spec-common/injectHeadless';
 import { getCLIHost, loadNativeModule } from '../spec-common/commonUtils';
 import { resolve } from './configContainer';
 import { URI } from 'vscode-uri';
-import { promisify } from 'util';
 import { LogLevel, LogDimensions, toErrorText, createCombinedLog, createTerminalLog, Log, makeLog, LogFormat, createJSONLog, createPlainLog } from '../spec-utils/log';
 import { dockerComposeCLIConfig } from './dockerCompose';
 import { Mount } from '../spec-configuration/containerFeaturesConfiguration';
@@ -103,7 +102,7 @@ export async function createDockerParams(options: ProvisionOptions, disposables:
 	const appRoot = undefined;
 	const cwd = options.workspaceFolder || process.cwd();
 	const cliHost = await getCLIHost(cwd, loadNativeModule);
-	const sessionId = (await promisify(crypto.randomBytes)(20)).toString('hex'); // TODO: Somehow enable correlation.
+	const sessionId = crypto.randomUUID();
 
 	const common: ResolverParameters = {
 		prebuild: options.prebuild,
@@ -123,7 +122,7 @@ export async function createDockerParams(options: ProvisionOptions, disposables:
 		output,
 		allowSystemConfigChange: true,
 		defaultUserEnvProbe: options.defaultUserEnvProbe,
-		postCreate: createNullPostCreate(options.postCreateEnabled, options.skipNonBlocking, output),
+		lifecycleHook: createNullLifecycleHook(options.postCreateEnabled, options.skipNonBlocking, output),
 		getLogLevel: () => options.logLevel,
 		onDidChangeLogLevel: () => ({ dispose() { } }),
 		loadNativeModule,