+++ src/test/container-features/configs/lifecycle-hooks-advanced/.devcontainer/rabbit/helper_script.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+MARKER_FILE_NAME="$1"
+echo "Hello from rabbit helper_script.sh invoked by ${MARKER_FILE_NAME}"
+touch "helperScript.rabbit.${MARKER_FILE_NAME}.testMarker"